requirements.txt
streamlit==1.28.0
streamlit-shadcn-ui==0.1.18
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
supabase==2.0.0
